+++ test/e2e_node/BUILD
@@ -38,6 +38,7 @@ go_library(
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
         "//staging/src/k8s.io/apiserver/pkg/util/feature:go_default_library",
+        "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
         "//staging/src/k8s.io/client-go/kubernetes/scheme:go_default_library",
         "//staging/src/k8s.io/component-base/featuregate:go_default_library",
         "//staging/src/k8s.io/cri-api/pkg/apis:go_default_library",
@@ -67,7 +68,6 @@ go_library(
             "//staging/src/k8s.io/apimachinery/pkg/util/runtime:go_default_library",
             "//staging/src/k8s.io/apimachinery/pkg/util/uuid:go_default_library",
             "//staging/src/k8s.io/apimachinery/pkg/util/wait:go_default_library",
-            "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
             "//staging/src/k8s.io/client-go/kubernetes/typed/core/v1:go_default_library",
             "//test/e2e/framework/kubelet:go_default_library",
             "//test/e2e/framework/perf:go_default_library",
